:sparkles: [Feat]: BroadcastPlayer 컴포넌트 생성
